requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
python-multipart
supabase
